Refactor: extract cache.py and watcher.py from indexer.py
indexer.py (811 行) を責務単位に分解:
- cache.py: CacheEntry / TieredCache (Tier 0/1 の in-memory cache)
- watcher.py: VaultWatcher (watchdog ベースのファイル監視)
- indexer.py: VaultIndex のみ残す (FTS5 + SQLite、Tier 2)

re-export は行わず、callsite (server.py / tests/test_indexer.py /
tests/test_watcher.py) の import パスを直接更新。Opus 評価が PEP 8 上の
後方互換 re-export を reflexive なバイアスとして指摘した (Phase B2 と同方針)。

254 tests green / ruff clean。挙動変更なし。
