+++ tests/test_indexer.py
@@ -7,7 +7,8 @@ from pathlib import Path
 
 import pytest
 
-from vault_search.indexer import TieredCache, VaultIndex
+from vault_search.cache import TieredCache
+from vault_search.indexer import VaultIndex
 from vault_search.validation import ValidationError
 
 # ---------------------------------------------------------------------------
